\name{objectfunctionsbooster-package}
\alias{objectfunctionsbooster-package}
\alias{objectfunctionsbooster}
\docType{package}
\title{
\packageTitle{objectfunctionsbooster}
}
\description{
\packageDescription{objectfunctionsbooster}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{objectfunctionsbooster}
\packageIndices{objectfunctionsbooster}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{objectfunctionsbooster}

Maintainer: \packageMaintainer{objectfunctionsbooster}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
